### 2019-02-21 #1270 @whedon
[PRE REVIEW]: tsfeaturex: An R Package for Automating Time Series Feature Extraction
**Submitting author:** @nelsonroque (<a href="http://orcid.org/0000-0003-1184-202X">Nelson Roque</a>)
**Repository:** <a href="https://github.com/nelsonroque/tsfeaturex" target ="_blank">https://github.com/nelsonroque/tsfeaturex</a>
**Version:** v.0.3.6
