### 2019-02-25 #1270 @xuanxu
👋  @acolum, @aj2duncan: Would you be willing to review this paper? I think your expertise in R and math makes both of you a great fit for this submission.
I think one reviewer is enough to get this started, but it would be great to have the two of you reviewing it.
