### 2019-02-26 #1270 @xuanxu
@acolum, @aj2duncan, please head over to #1279 for the actual review
