### 2019-03-03 #1279 @aj2duncan
Hi @nelsonroque, I'm having a little trouble with the example that you've got in the README.md of your repo. When I run

``` r
library(tsfeaturex)

# for reproducibility of this example
set.seed(516)

# create test data
dat <- data.frame(expand.grid(day=c(1:7),id=c(1:100)))
dat$y <- rnorm(nrow(dat),5,1.5)
dat$y[1:3] <- NA # introduce NAs to check

# run function
out.list <- extract_features(df=dat,group_var="id",value_var="y",features="all")
```
I get the error
```
 Error in packageVersion("featuRe") : package ‘featuRe’ not found
```

I will keep working on my checklist but wanted to flag this up just now.
