+++ NAMESPACE
@@ -88,7 +88,6 @@ export(f.which.below.3sigma)
 export(feature_corr_heatmap)
 export(feature_correlations)
 export(features_to_df)
-export(extract_features)
 export(generate_extraction_list)
 export(normalize)
 import(entropy)
