+++ NAMESPACE
@@ -85,6 +85,7 @@ export(f.which.above.3sigma)
 export(f.which.below.1sigma)
 export(f.which.below.2sigma)
 export(f.which.below.3sigma)
+export(extract_features)
 export(feature_corr_heatmap)
 export(feature_correlations)
 export(features_to_df)
