+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(extract_features)
 export(f.IQR)
 export(f.autocorrelation)
 export(f.autocorrelation.lag1.1st.val)
@@ -85,7 +86,6 @@ export(f.which.above.3sigma)
 export(f.which.below.1sigma)
 export(f.which.below.2sigma)
 export(f.which.below.3sigma)
-export(extract_features)
 export(feature_corr_heatmap)
 export(feature_correlations)
 export(features_to_df)
@@ -98,4 +98,87 @@ import(zoo)
 importFrom(Langevin,Langevin1D)
 importFrom(e1071,kurtosis)
 importFrom(e1071,skewness)
+importFrom(psych,"(add")
+importFrom(psych,"NAs)")
+importFrom(psych,"for")
+importFrom(psych,"function")
+importFrom(psych,"if")
+importFrom(psych,"in")
+importFrom(psych,"name,")
+importFrom(psych,"otherwise,")
+importFrom(psych,"provided,")
+importFrom(psych,a)
+importFrom(psych,acknowledgement)
+importFrom(psych,all)
+importFrom(psych,argument)
+importFrom(psych,as)
+importFrom(psych,assuming)
+importFrom(psych,based)
+importFrom(psych,bind)
+importFrom(psych,by)
+importFrom(psych,calculate)
+importFrom(psych,calculation)
+importFrom(psych,call)
+importFrom(psych,category)
+importFrom(psych,check)
+importFrom(psych,common)
+importFrom(psych,construct)
+importFrom(psych,create)
+importFrom(psych,custom)
+importFrom(psych,data)
+importFrom(psych,dataframe)
+importFrom(psych,dataset)
+importFrom(psych,desired)
+importFrom(psych,details)
+importFrom(psych,did)
+importFrom(psych,downstream)
+importFrom(psych,each)
+importFrom(psych,elapsed)
+importFrom(psych,end)
+importFrom(psych,export)
+importFrom(psych,extract)
+importFrom(psych,extraction)
+importFrom(psych,feature)
+importFrom(psych,features)
+importFrom(psych,first)
+importFrom(psych,from)
+importFrom(psych,full)
+importFrom(psych,get)
+importFrom(psych,into)
+importFrom(psych,list)
+importFrom(psych,method)
 importFrom(psych,mssd)
+importFrom(psych,multiple)
+importFrom(psych,name)
+importFrom(psych,namespace)
+importFrom(psych,not)
+importFrom(psych,object)
+importFrom(psych,of)
+importFrom(psych,on)
+importFrom(psych,one)
+importFrom(psych,parameter)
+importFrom(psych,pass)
+importFrom(psych,print)
+importFrom(psych,process)
+importFrom(psych,provided)
+importFrom(psych,purposes)
+importFrom(psych,request)
+importFrom(psych,return)
+importFrom(psych,run)
+importFrom(psych,save)
+importFrom(psych,start)
+importFrom(psych,store)
+importFrom(psych,storing)
+importFrom(psych,support)
+importFrom(psych,the)
+importFrom(psych,they)
+importFrom(psych,time)
+importFrom(psych,timer)
+importFrom(psych,timing)
+importFrom(psych,timings)
+importFrom(psych,to)
+importFrom(psych,user)
+importFrom(psych,variable)
+importFrom(psych,vector)
+importFrom(psych,view)
+importFrom(psych,with)
