+++ changelog.md
@@ -1,0 +1,7 @@
+# Changelog
+
+### version 0.3.8
+-- added convenience function `plot_series` to quickly plot a time-series for one or many data streams.
+
+### version 0.3.7
+-- patched long-to-wide issues with features_to_df()
