+++ R/extract_features.R
@@ -11,13 +11,13 @@
 #' @examples
 #' extract_features(df,group_var='id',value_var='y',features='simple')
 #' extract_features(df,group_var=c('id','month'),'y',features='all')
-#' @export 
 #' @import entropy
 #' @import zoo
 #' @importFrom e1071 kurtosis skewness
 #' @importFrom Langevin Langevin1D
 #' @importFrom psych mssd
 
+#' @export
 extract_features <- function(df, group_var, value_var, features='all', custom_feature_list = NA, data_confirm=T, return_timing=T, verbose=T) {
   VERSION_CODE = packageVersion("tsfeaturex")   #' storing for multiple downstream purposes
   
