+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### version 0.3.9
+-- added warnings based on PSU QuantDev feedback (3/27/19; tsfeaturex demo + presentation)
+
 ### version 0.3.8
 -- added convenience function `plot_series` to quickly plot a time-series for one or many data streams.
 
