+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: tsfeaturex
-Version: 0.3.8
+Version: 0.3.9
 Date: 2019-03-28
 Title: Calculates Numerous Features For Any Time Series
 Authors@R: person("Nelson", "Roque", email = "nelsonroquejr@gmail.com", role = c("aut", "cre"))
