+++ R/extract_features.R
@@ -8,18 +8,19 @@
 #' @param data_confirm class: boolean; T = confirming data is in long format
 #' @param return_timing class: boolean; T = return dataframe with timing for each feature
 #' @param verbose class: boolean; T = display status, debugging info
-#' @export
 #' @examples
 #' extract_features(df,group_var='id',value_var='y',features='simple')
 #' extract_features(df,group_var=c('id','month'),'y',features='all')
+#' @export
 #' @import entropy
 #' @import zoo
-#' @importFrom psych mssd
 #' @importFrom e1071 kurtosis skewness
 #' @importFrom Langevin Langevin1D
+#' @importFrom psych mssd
 
 extract_features <- function(df, group_var, value_var, features='all', custom_feature_list=NA, data_confirm=T, return_timing=T, verbose=T){
-  VERSION_CODE = packageVersion("tsfeaturex")   #' storing for multiple downstream purposes
+  VERSION_CODE = packageVersion("tsfeaturex")   
+  #' storing for multiple downstream purposes
   
   if(!data_confirm){
     #' request acknowledgement from user if they did not pass in the parameter
@@ -57,24 +58,16 @@ extract_features <- function(df, group_var, value_var, features='all', custom_fe
     SESSION_START_TIME <- Sys.time()
     SESSION_START_TIMESTAMP = format(SESSION_START_TIME, "%m_%d_%Y-%H_%M_%S")
     
-    #' 
-    
     #' process all features
     for(feature_ in current.features) {
-  
-      #' 
-      
+        
       #' start feature timer
       PROCESS_START_TIME = Sys.time()
       PROCESS_START_TIMESTAMP = format(PROCESS_START_TIME, "%m/%d/%Y %H:%M:%S")
-      
-      #' 
-      
+            
       #' extract function name, assuming first argument
       func_name <- feature_[1]
-      
-      #' 
-      
+            
       #' construct function call
       if(length(feature_) > 1) {
         func_var_str <- paste(func_name,func_args,sep='_')
@@ -85,48 +78,32 @@ extract_features <- function(df, group_var, value_var, features='all', custom_fe
         func_var_str <- paste(func_name,sep='_')
         method_constructor <- paste(func_name,'(',value_var,')')
       }
-      
-      #' 
-      
-      #' construct feature method (add in support for NAs)
+            
+      #' construct feature method add in support for NAs
       new_method <- method_constructor
-      
-      #' 
-      
+            
       #' create variable name
       new_var_name <- paste0(func_var_str,'_', value_var)
-      
-      #' 
-      
+            
       #' start feature extraction timer
       FEATURE_START_TIME = Sys.time()
       FEATURE_START_TIMESTAMP = format(FEATURE_START_TIME, "%m/%d/%Y %H:%M:%S")
-      
-      #' 
-      
+            
       #' run one feature calculation
       df_output <- df %>%
         group_by_(.dots = group_var) %>%
         summarise_(.dots = setNames(new_method, new_var_name))
-      
-      #' 
-      
+            
       #' save each feature as list object
       feature_list[[feature_]] <- df_output
-      
-      #' 
-      
+            
       #' end feature timer
       PROCESS_END_TIME = Sys.time()
-
-      #' 
       
       #' calculate elapsed time
       feature_elapsed_time = difftime(PROCESS_END_TIME,FEATURE_START_TIME,units='secs')
       process_elapsed_time = difftime(PROCESS_END_TIME,PROCESS_START_TIME,units='secs')
-      
-      #' 
-      
+            
       #' create export dataset with timing features
       feature_export_df <- data.frame(feature = func_name,
[…]
