+++ NAMESPACE
@@ -1,6 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export(extract_features)
 export(f.IQR)
 export(f.autocorrelation)
 export(f.autocorrelation.lag1.1st.val)
@@ -95,87 +94,9 @@ export(plot_series)
 import(entropy)
 import(reshape2)
 import(tidyverse)
+import(viridis)
 import(zoo)
-importFrom(Langevin,"(add")
-importFrom(Langevin,"NAs)")
-importFrom(Langevin,"for")
-importFrom(Langevin,"function")
-importFrom(Langevin,"if")
-importFrom(Langevin,"in")
-importFrom(Langevin,"name,")
-importFrom(Langevin,"otherwise,")
-importFrom(Langevin,"provided,")
 importFrom(Langevin,Langevin1D)
-importFrom(Langevin,a)
-importFrom(Langevin,acknowledgement)
-importFrom(Langevin,all)
-importFrom(Langevin,argument)
-importFrom(Langevin,as)
-importFrom(Langevin,assuming)
-importFrom(Langevin,based)
-importFrom(Langevin,bind)
-importFrom(Langevin,by)
-importFrom(Langevin,calculate)
-importFrom(Langevin,calculation)
-importFrom(Langevin,call)
-importFrom(Langevin,category)
-importFrom(Langevin,check)
-importFrom(Langevin,common)
-importFrom(Langevin,construct)
-importFrom(Langevin,create)
-importFrom(Langevin,custom)
-importFrom(Langevin,data)
-importFrom(Langevin,dataframe)
-importFrom(Langevin,dataset)
-importFrom(Langevin,desired)
-importFrom(Langevin,details)
-importFrom(Langevin,did)
-importFrom(Langevin,each)
-importFrom(Langevin,elapsed)
-importFrom(Langevin,end)
-importFrom(Langevin,export)
-importFrom(Langevin,extract)
-importFrom(Langevin,extraction)
-importFrom(Langevin,feature)
-importFrom(Langevin,features)
-importFrom(Langevin,first)
-importFrom(Langevin,from)
-importFrom(Langevin,full)
-importFrom(Langevin,get)
-importFrom(Langevin,into)
-importFrom(Langevin,list)
-importFrom(Langevin,method)
-importFrom(Langevin,name)
-importFrom(Langevin,namespace)
-importFrom(Langevin,not)
-importFrom(Langevin,object)
-importFrom(Langevin,of)
-importFrom(Langevin,on)
-importFrom(Langevin,one)
-importFrom(Langevin,parameter)
-importFrom(Langevin,pass)
-importFrom(Langevin,print)
-importFrom(Langevin,process)
-importFrom(Langevin,provided)
-importFrom(Langevin,request)
-importFrom(Langevin,return)
-importFrom(Langevin,run)
-importFrom(Langevin,save)
-importFrom(Langevin,start)
-importFrom(Langevin,store)
-importFrom(Langevin,support)
-importFrom(Langevin,the)
-importFrom(Langevin,they)
-importFrom(Langevin,time)
-importFrom(Langevin,timer)
-importFrom(Langevin,timing)
-importFrom(Langevin,timings)
-importFrom(Langevin,to)
-importFrom(Langevin,user)
-importFrom(Langevin,variable)
-importFrom(Langevin,vector)
-importFrom(Langevin,view)
-importFrom(Langevin,with)
 importFrom(e1071,kurtosis)
 importFrom(e1071,skewness)
 importFrom(psych,mssd)
