+++ NAMESPACE
@@ -92,93 +92,11 @@ export(features_to_df)
 export(generate_extraction_list)
 export(normalize)
 export(plot_series)
-import("for")
-import("function")
-import("if")
-import("in")
-import("name,")
-import("otherwise,")
-import("provided,")
-import(NAs)
-import(a)
-import(acknowledgement)
-import(add)
-import(all)
-import(argument)
-import(as)
-import(assuming)
-import(based)
-import(bind)
-import(by)
-import(calculate)
-import(calculation)
-import(call)
-import(category)
-import(check)
-import(common)
-import(construct)
-import(create)
-import(custom)
-import(data)
-import(dataframe)
-import(dataset)
-import(desired)
-import(details)
-import(did)
-import(downstream)
-import(each)
-import(elapsed)
-import(end)
 import(entropy)
-import(export)
-import(extract)
-import(extraction)
-import(feature)
-import(features)
-import(first)
-import(from)
-import(full)
-import(get)
-import(into)
-import(list)
-import(method)
-import(multiple)
-import(name)
-import(namespace)
-import(not)
-import(object)
-import(of)
-import(on)
-import(one)
-import(parameter)
-import(pass)
-import(print)
-import(process)
-import(provided)
-import(purposes)
-import(request)
 import(reshape2)
-import(return)
-import(run)
-import(save)
-import(start)
-import(store)
-import(storing)
-import(support)
-import(the)
-import(they)
 import(tidyverse)
-import(time)
-import(timer)
-import(timing)
-import(timings)
-import(to)
-import(user)
-import(variable)
-import(vector)
-import(view)
+import(tidyr)
 import(viridis)
-import(with)
 import(zoo)
 importFrom(Langevin,Langevin1D)
 importFrom(e1071,kurtosis)
