+++ NAMESPACE
@@ -95,7 +95,6 @@ export(plot_series)
 import(entropy)
 import(reshape2)
 import(tidyverse)
-import(tidyr)
 import(viridis)
 import(zoo)
 importFrom(Langevin,Langevin1D)
