### 2019-04-24 #1279 @aj2duncan
There has actually been progress made, in that having noticed the package has been updated, I reinstalled it and reran the example code. This time I get a different error.

The example code currently reads

```{r}
# load library
library(tsfeaturex)

# for reproducibility of this example
set.seed(516)

# create test data
dat <- data.frame(expand.grid(day=c(1:7),id=c(1:100)))
dat$y <- rnorm(nrow(dat),5,1.5)
dat$y[1:3] <- NA # introduce NAs to check

# run function
out.list <- extract_features(df=dat,group_var="id",value_var="y",features="all")

# convert list to data.frame (MapReduce)
final.df <- features_to_df(out.list, data.format="wide")

# get feature correlations
cor.df <- feature_correlations(final.df, data.format="wide")

# view results
View(final.df)
```
This gives errors about `group_var` and `id_var` within the functions `final.df()` and `cor.d()`. However, if I add these so the code becomes

```{r}
# load library
library(tsfeaturex)

# for reproducibility of this example
set.seed(516)

# create test data
dat <- data.frame(expand.grid(day=c(1:7),id=c(1:100)))
dat$y <- rnorm(nrow(dat),5,1.5)
dat$y[1:3] <- NA # introduce NAs to check

# run function
out.list <- extract_features(df=dat,group_var="id",value_var="y",features="all")

# convert list to data.frame (MapReduce)
final.df <- features_to_df(out.list, data.format="wide", group_var = "id")

# get feature correlations
cor.df <- feature_correlations(final.df, data.format="wide", id_var = "id")

# view results
View(final.df)
```
The code runs and I only get a warning of:
```
Warning message:
In cor(.) : the standard deviation is zero
```
As the package is now working I will continue with the review. @nelsonroque could you please update the example code so it runs without error.
