### 2019-05-01 #1279 @nelsonroque
@labarba @aj2duncan @arfon the example code has been modified to run without error. Thank you for your time and feedback!
