### 2019-05-11 #1279 @acolum
@nelsonroque I second @aj2duncan's suggestions and can add a few more.

In the paper under the Figures section, I think you should elaborate more on your example. It would be beneficial to see an example (with code) of how tsfeaturex can calculate different features and characterize differences between two time series with the same mean but different shapes/peaks. I also think that you should clarify with a diagram and/or more detailed description what long and wide format data means for this package. This would benefit both specialist and non-specialist users.

In the documentation, the dependencies in the description file should also be added to the readme.
