### 2019-05-14 988e96a
adding information suggested by reviewers on JOSS
