### 2019-05-14 26a057d
minor formatting around new changes
