### 2019-05-14 f83827c
updated formatting of feature list
