+++ feature_list.md
@@ -1,83 +1,83 @@
 # Simple Features
-min
-mean
-median
-max
-sum
-IQR
-sd
-variance
-coeff.variation
-range
+- min
+- mean
+- median
+- max
+- sum
+- IQR
+- sd
+- variance
+- coeff.variation
+- range
 
 # Binary Features
-sd.greater.range
+- sd.greater.range
 
 # Distribution Features
-kurtosis
-skew
-quantile_10
-quantile_25
-quantile_30
-quantile_40
-quantile_50
-quantile_60
-quantile_70
-quantile_75
-quantile_80
-quantile_90
-quantile_95
-count.above.mean
-count.below.mean
-count.negative
-count.positive
-count_NA
-length
-perc.missing
-prop.missing
-ratio.unique.values
-perc.unique.values
-autocorrelation.lag1.string
-autocorrelation.lag1.1st.val
-autocorrelation.lag1.2nd.val
-autocorrelation.lag2.string
-autocorrelation.lag2.1st.val
-autocorrelation.lag2.2nd.val
-partial.autocorrelation.lag1.string
-partial.autocorrelation.lag1.1st.val
-partial.autocorrelation.lag1.2nd.val
-partial.autocorrelation.lag2.string
-partial.autocorrelation.lag2.1st.val
-partial.autocorrelation.lag2.2nd.val
-entropy
-mssd.lag1
-mssd.lag2
-mssd.lag3
-mean.change
-median.change
-min.change
-max.change
-sum.change
-prob.acute.change.1
-prob.acute.change.2
-prob.acute.change.3
-prob.acute.change.4
-prob.acute.change.5
-prob.acute.change.6
-prob.acute.change.7
-prob.acute.change.8
-prob.acute.change.9
-prob.acute.change.95
-prob.acute.change.99
-count.below.1sigma
-which.below.1sigma
-count.above.1sigma
-which.above.1sigma
-count.below.2sigma
-which.below.2sigma
-count.above.2sigma
-which.above.2sigma
-count.below.3sigma
-which.below.3sigma
-count.above.3sigma
-which.above.3sigma
+- kurtosis
+- skew
+- quantile_10
+- quantile_25
+- quantile_30
+- quantile_40
+- quantile_50
+- quantile_60
+- quantile_70
+- quantile_75
+- quantile_80
+- quantile_90
+- quantile_95
+- count.above.mean
+- count.below.mean
+- count.negative
+- count.positive
+- count_NA
+- length
+- perc.missing
+- prop.missing
+- ratio.unique.values
+- perc.unique.values
+- autocorrelation.lag1.string
+- autocorrelation.lag1.1st.val
+- autocorrelation.lag1.2nd.val
+- autocorrelation.lag2.string
+- autocorrelation.lag2.1st.val
+- autocorrelation.lag2.2nd.val
+- partial.autocorrelation.lag1.string
+- partial.autocorrelation.lag1.1st.val
+- partial.autocorrelation.lag1.2nd.val
+- partial.autocorrelation.lag2.string
+- partial.autocorrelation.lag2.1st.val
+- partial.autocorrelation.lag2.2nd.val
+- entropy
+- mssd.lag1
+- mssd.lag2
+- mssd.lag3
+- mean.change
+- median.change
+- min.change
+- max.change
+- sum.change
+- prob.acute.change.1
+- prob.acute.change.2
+- prob.acute.change.3
+- prob.acute.change.4
+- prob.acute.change.5
+- prob.acute.change.6
+- prob.acute.change.7
+- prob.acute.change.8
+- prob.acute.change.9
+- prob.acute.change.95
+- prob.acute.change.99
+- count.below.1sigma
+- which.below.1sigma
[…]
