### 2019-05-14 cbcf6ac
initial ideas for unit tests
