### 2019-05-20 9c96ad6
added auto testing
will expand over time
