+++ DESCRIPTION
@@ -11,3 +11,5 @@ URL: https://github.com/nelsonroque/featuRe
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
+Suggests: 
+    testthat (>= 2.1.0)
