+++ paper/paper.md
@@ -45,6 +45,10 @@ Short data streams are easily summarized using basic features (e.g., mean, stand
 
 For convenience, the package can export extracted features in long format or wide format (i.e., wide format may be of interest when attempting to correlate same feature across several variables of interest).  By design, `tsfeaturex` is able to cope with missing data (in R, of class `NA`), a key deviation from tsfresh [@tsfreshgitissue].  In addition to feature extraction, this package also calculates feature correlations amongst outputted features.
 
+### Data Structures
+
+`tsfeaturex` is capable of outputting both `long` and `wide` data structures -- both of use for different purposes (e.g., `long` format preferred for plotting in ggplot2) and analyses (e.g., `wide` format preferred for repeated measures ANOVA in most statistical software).  Figure 1 depicts example `long` and `wide` data structures. ![Flexible data structure output -- request `long` or `wide` format](data_shape.png)
+
 ## Purpose & Audience
 `tsfeaturex` is intended for use by researchers with time-series data, and will be of most interest to those developing their statistical and coding skills -- allowing them to extract many features from their time-series data with easy to use code and without need for high-level mathematics background.  The desire for feature extraction tools is widespread across all domains of data science, including, but not limited to, applications in: biological systems, finance, and psychology.
 
@@ -64,7 +68,7 @@ The current expectation is that over time, `tsfeaturex`, will allow for two-leve
     * [Learn More About This Project](https://sliwinskilab.weebly.com/research-projects.html)
 
 # Figures
-Figure depicts sample time series data from two participants, both with mean value of 5. You will notice, although they have identical means, the shape of the time series, and locations of peaks is different. `tsfeaturex` calculates features to better characterize differences such as these. ![The mean doesn't tell the whole story](figure.png)
+Figure 2 depicts sample time series data from two participants, both with mean value of 5. You will notice, although they have identical means, the shape of the time series, and locations of peaks is different. `tsfeaturex` calculates features to better characterize differences such as these. ![The mean doesn't tell the whole story](figure.png)
 
 # Acknowledgements
 Nelson A. Roque was supported by National Institute on Aging Grant T32 AG049676 to The Pennsylvania State University.
