+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### version 0.3.10 | Date: 5/20/19
+-- added automated tests (using: 'testthat')
+-- updated readme.md documentation to include more information for support
+-- added statement of need, and figures for visual presetnation of data structures (long, wide)
+
 ### version 0.3.9
 -- added warnings based on PSU QuantDev feedback (3/27/19; tsfeaturex demo + presentation)
 
