+++ paper/paper.bib
@@ -7,15 +7,18 @@ title = {{10 Key Marketing Trends for 2017 and Ideas for Exceeding Customer Expe
 url = {https://www.ibm.com/downloads/cas/XKBEABLN}
 }
 
-
 @article{tsfresh,
-  title={Time series feature extraction on basis of scalable hypothesis tests (tsfresh--a python package)},
-  author={Christ, Maximilian and Braun, Nils and Neuffer, Julius and Kempa-Liehr, Andreas W},
-  journal={Neurocomputing},
-  volume={307},
-  pages={72--77},
-  year={2018},
-  publisher={Elsevier}
+  title = "Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests (tsfresh: A Python package)",
+  journal = "Neurocomputing",
+  volume = "307",
+  pages = "72 - 77",
+  year = "2018",
+  issn = "0925-2312",
+  doi = "https://doi.org/10.1016/j.neucom.2018.03.067",
+  url = "http://www.sciencedirect.com/science/article/pii/S0925231218304843",
+  author = "Maximilian Christ and Nils Braun and Julius Neuffer and Andreas W. Kempa-Liehr",
+  keywords = "Feature engineering, Time series, Feature extraction, Feature selection, Machine learning",
+  abstract = "Time series feature engineering is a time-consuming process because scientists and engineers have to consider the multifarious algorithms of signal processing and time series analysis for identifying and extracting meaningful features from time series. The Python package tsfresh (Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests) accelerates this process by combining 63 time series characterization methods, which by default compute a total of 794 time series features, with feature selection on basis automatically configured hypothesis tests. By identifying statistically significant time series characteristics in an early stage of the data science process, tsfresh closes feedback loops with domain experts and fosters the development of domain specific features early on. The package implements standard APIs of time series and machine learning libraries (e.g. pandas and scikit-learn) and is designed for both exploratory analyses as well as straightforward integration into operational data science applications."
 }
 
 @misc{tsfreshgit,
@@ -71,3 +74,20 @@ year = {2008}
   doi = {10.5281/zenodo.2574990},
   url = {http://dx.doi.org/10.5281/zenodo.2574990}
 }
+
+@Manual{r_cite,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2019},
+  url = {https://www.R-project.org/},
+}
+
+@Manual{r_shiny,
+    title = {shiny: Web Application Framework for R},
+    author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
+    year = {2019},
+    note = {R package version 1.3.1},
+    url = {https://CRAN.R-project.org/package=shiny},
+  }
