### 2019-05-20 d96d18b
updated citation links for packages
