+++ paper/paper.bib
@@ -91,3 +91,105 @@ year = {2008}
     note = {R package version 1.3.1},
     url = {https://CRAN.R-project.org/package=shiny},
   }
+  
+@Manual{r_psych,
+    title = {psych: Procedures for Psychological, Psychometric, and Personality Research},
+    author = {William Revelle},
+    organization = { Northwestern University},
+    address = { Evanston, Illinois},
+    year = {2018},
+    note = {R package version 1.8.12},
+    url = {https://CRAN.R-project.org/package=psych},
+  }
+  
+  
+ @Manual{r_e1071,
+    title = {e1071: Misc Functions of the Department of Statistics, Probability
+Theory Group (Formerly: E1071), TU Wien},
+    author = {David Meyer and Evgenia Dimitriadou and Kurt Hornik and Andreas Weingessel and Friedrich Leisch},
+    year = {2019},
+    note = {R package version 1.7-1},
+    url = {https://CRAN.R-project.org/package=e1071},
+  }
+  
+ @Manual{r_entropy,
+    title = {entropy: Estimation of Entropy, Mutual Information and Related Quantities},
+    author = {Jean Hausser and Korbinian Strimmer},
+    year = {2014},
+    note = {R package version 1.2.1},
+    url = {https://CRAN.R-project.org/package=entropy},
+  }
+  
+  @Article{r_forecast,
+  title = {Automatic time series forecasting: the forecast package for {R}},
+  author = {Rob J Hyndman and Yeasmin Khandakar},
+  journal = {Journal of Statistical Software},
+  volume = {26},
+  number = {3},
+  pages = {1--22},
+  year = {2008},
+  url = {http://www.jstatsoft.org/article/view/v027i03},
+}
+
+
+@Article{Rinn2016,
+  title = {The Langevin Approach: An R Package for Modeling Markov Processes},
+  author = {Philip Rinn and Pedro G. Lind and Matthias WÃ¤chter and Joachim Peinke},
+  journal = {Journal of Open Research Software},
+  year = {2016},
+  number = {1},
+  pages = {e34},
+  volume = {4},
+  doi = {10.5334/jors.123},
+  issn = {2049-9647},
+  publisher = {Ubiquity Press},
+}
+
+@Article{Friedrich2011,
+  title = {Approaching complexity by stochastic methods: From biological systems to turbulence},
+  author = {Rudolf Friedrich and Joachim Peinke and Muhammad Sahimi and Mohammed {Reza Rahimi Tabar}},
+  journal = {Physics Reports},
+  year = {2011},
+  number = {5},
+  pages = {87--162},
+  volume = {506},
+  doi = {10.1016/j.physrep.2011.05.003},
+  issn = {0370-1573},
+  publisher = {Elsevier BV},
+}
+
+@Manual{r_viridis,
+  title = {viridis: Default Color Maps from 'matplotlib'},
+  author = {Simon Garnier},
+  year = {2018},
+  note = {R package version 0.5.1},
+  url = {https://CRAN.R-project.org/package=viridis},
+}
+
+
+@Article{r_zoo,
+  title = {zoo: S3 Infrastructure for Regular and Irregular Time Series},
+  author = {Achim Zeileis and Gabor Grothendieck},
+  journal = {Journal of Statistical Software},
+  year = {2005},
+  volume = {14},
+  number = {6},
+  pages = {1--27},
+  doi = {10.18637/jss.v014.i06},
+}
+
+@Manual{r_hmisc,
+  title = {Hmisc: Harrell Miscellaneous},
+  author = {Frank E {Harrell Jr} and with contributions from Charles Dupont and many others.},
+  year = {2019},
[…]
