+++ paper/paper.md
@@ -42,7 +42,7 @@ Short data streams are easily summarized using basic features (e.g., mean, stand
 # Summary
 
 ## Functionality
-`tsfeaturex` [@roque_ram_2019] is an R package for automating time series feature extraction, inspired and modeled after the Python package *tsfresh* [@tsfresh; @tsfreshgit].  The R language [@r_cite] allows for an easy to use interface, with the underlying processing speed advantage of C languages (and flexibility to run on the web, with the help of the `shiny` package in R; [@r_shiny]).  The API for `tsfeaturex` is a wrapper for the highly-trafficked 'dplyr', mainly to lend on the flexibility of the grammar of data manipulation and shortcuts for non-standard evaluation. The API for `tsfeaturex` was designed to facilitate the extraction of features for any dataset in long format, including grouping of summaries by other factor. For example, if every person in your dataset has 1 observation each day for 8 days, and they do this in two bursts, once every 6 months, you can calculate features of the overall series, 16 observations from both bursts, or separately for each burst).
+`tsfeaturex` [@roque_ram_2019] is an R package for automating time series feature extraction, inspired and modeled after the Python package *tsfresh* [@tsfresh; @tsfreshgit].  The R language [@r_cite] allows for an easy to use interface, with the underlying processing speed advantage of C languages (and flexibility to run on the web, with the help of the `shiny` package in R; [@r_shiny]).  The API for `tsfeaturex` is a wrapper for the highly-trafficked `dplyr`[@r_dplyr], mainly to lend on the flexibility of the grammar of data manipulation and shortcuts for non-standard evaluation. The API for `tsfeaturex` was designed to facilitate the extraction of features for any dataset in long format, including grouping of summaries by other factor. For example, if every person in your dataset has 1 observation each day for 8 days, and they do this in two bursts, once every 6 months, you can calculate features of the overall series, 16 observations from both bursts, or separately for each burst). Some features are integrated from other packages, such as: [`e1071`, @r_e1071; `Hmisc`, @r_hmisc; `forecast`, @r_forecast; `zoo`, @r_zoo; `viridis`, @r_viridis; `psych`, @r_psych; `entropy`, @r_entropy; `Langevin`, @Friedrich2011; @Rinn2016].
 
 By design, `tsfeaturex` is able to cope with missing data (in R, of class `NA`), a key deviation from tsfresh [@tsfreshgitissue].  In addition to feature extraction, this package also calculates feature correlations amongst outputted features.
 
@@ -67,9 +67,9 @@ The current expectation is that over time, `tsfeaturex`, will allow for two-leve
     * [Learn More About This Project](https://sliwinskilab.weebly.com/research-projects.html)
 
 # Figures
-Figure 1 depicts example `wide`(top) and `long`(bottom) data structures for a dataset containing two (2) measurements from two (2) individuals.  Notice that there is one row for each individual in the `wide` format, and two (2) rows for each individual in the `long` format, one for each column. ![Figure 1. Flexible data structure output -- request `long` or `wide` format](datashape.PNG)
+Figure 1 depicts example `wide`(top) and `long`(bottom) data structures for a dataset containing two (2) measurements from two (2) individuals.  Notice that there is one row for each individual in the `wide` format, and two (2) rows for each individual in the `long` format, one for each column. ![Figure 1. Flexible data structure output -- request `long` or `wide` format](datashape.PNG "Figure 1. Flexible data structure output -- request `long` or `wide` format")
 
[…]
