+++ paper/paper.bib
@@ -1,10 +1,9 @@
 @techreport{ibmstats,
-author = {{IBM Marketing Cloud}},
-file = {:C$\backslash$:/Users/nur375/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Unknown - Unknown - 0 Key Marketing Trends for 2017 and Ideas for Exceeding Customer Expectations.pdf:pdf},
-institution = {IBM},
-year={2017},
-title = {{10 Key Marketing Trends for 2017 and Ideas for Exceeding Customer Expectations}},
-url = {https://www.ibm.com/downloads/cas/XKBEABLN}
+  author = {IBM Marketing Cloud},
+  institution = {IBM},
+  year={2017},
+  title = {{10 Key Marketing Trends for 2017 and Ideas for Exceeding Customer Expectations}},
+  url = {https://www.ibm.com/downloads/cas/XKBEABLN}
 }
 
 @article{tsfresh,
@@ -49,19 +48,19 @@ url = {https://www.ibm.com/downloads/cas/XKBEABLN}
 }
 
 @article{Jahng2008,
-abstract = {Temporal instability of affect is a defining characteristic of psychological disorders such as borderline personality disorder (BPD) and mood cycling disorders. Ecological momentary assessment (EMA) enables researchers to directly assess such frequent and extreme fluctuations over time. The authors examined 4 operationalizations of such temporal instability: the within-person variance (WPV), the first-order autocorrelation, the mean square successive difference (MSSD), and the probability of acute change (PAC). It is argued that the MSSD and PAC measures are preferred indices of affective instability because they capture both variability and temporal dependency in a time series. Additionally, the performance of these 2 measures in capturing within- and between-day instability is discussed. To illustrate, the authors present EMA data from a study of negative mood in BPD and major depressive disorder patients. In this study, MSSD and PAC captured affective instability better than did WPV. Given that MSSD and PAC are individual difference measures, the authors propose that group differences on these indices be explored using generalized multilevel models. Versions of MSSD and PAC that adjust for randomly elapsed time interval between assessments are also presented.},
-author = {Jahng, Seungmin and Wood, Phillip K. and Trull, Timothy J.},
-doi = {10.1037/a0014173},
-issn = {1939-1463},
-journal = {Psychological Methods},
-month = {dec},
-number = {4},
-pages = {354--375},
-pmid = {19071999},
-title = {{Analysis of affective instability in ecological momentary assessment: Indices using successive difference and group comparison via multilevel modeling.}},
-url = {http://www.ncbi.nlm.nih.gov/pubmed/19071999 http://doi.apa.org/getdoi.cfm?doi=10.1037/a0014173},
-volume = {13},
-year = {2008}
[…]
