+++ paper/paper.md
@@ -7,7 +7,7 @@ authors:
 - affiliation: 2
   name: Dr. Nilam Ram
   orcid: 0000-0003-1671-5257
-date: "18 February 2019"
+date: "20 May 2019"
 output:
   html_document:
     df_print: paged
