+++ DESCRIPTION
@@ -7,7 +7,7 @@ Description: Inspired by tsfresh, this package calculate numerous features on an
 Imports: e1071, entropy, Langevin, Hmisc, forecast, stats, zoo, viridis, psych
 Depends: R (>= 3.4.0), tidyverse
 License: MIT
-URL: https://github.com/nelsonroque/featuRe
+URL: https://github.com/nelsonroque/tsfeaturex
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
