+++ docs/changelog.md
@@ -1,15 +1,16 @@
 # Changelog
 
-### version 0.3.10 | Date: 5/20/19
--- added automated tests (using: 'testthat')
--- updated readme.md documentation to include more information for support
--- added statement of need, and figures for visual presetnation of data structures (long, wide)
+## version 0.3.10
+### Date: 5/20/19
+- added automated tests (using: 'testthat')
+- updated readme.md documentation to include more information for support
+- added statement of need, and figures for visual presetnation of data structures (long, wide)
 
-### version 0.3.9
--- added warnings based on PSU QuantDev feedback (3/27/19; tsfeaturex demo + presentation)
+## version 0.3.9
+- added warnings based on PSU QuantDev feedback (3/27/19; tsfeaturex demo + presentation)
 
-### version 0.3.8
--- added convenience function `plot_series` to quickly plot a time-series for one or many data streams.
+## version 0.3.8
+- added convenience function `plot_series` to quickly plot a time-series for one or many data streams.
 
-### version 0.3.7
--- patched long-to-wide issues with features_to_df()
+## version 0.3.7
+- patched long-to-wide issues with features_to_df()
