### 2019-05-20 #1279 @nelsonroque
@aj2duncan @acolum thank you for all your feedback. I have addressed all points raised, including a new draft of the paper, with updated references, additions and changes to readme.md, a new docs/ folder on root with changelog and feature list in markdown format.

@acolum I had a question regarding your comment: "It would be beneficial to see an example (with code) of how tsfeaturex can calculate different features and characterize differences between two time series with the same mean but different shapes/peaks." I can easily simulate multiple data streams to have a common mean, and then calculate all features for each 'person' -- but how would you like me to visualize this? A graph like the one that exists now, with perhaps a data table underneath, showing each feature (currently would be 82 feature columns), and highlighting cells that are different?
