### 2019-05-28 #1279 @acolum
@nelsonroque After more thought, I've realized that including the example in my comment wouldn't be necessary given all of the other detailed documentation you've included.

Overall, I really liked all of the changes you've recently made to your package, documentation, and paper. With these improvements, I was able to check off everything on my review list and am recommending this paper for publication.
