### 2019-05-28 #1279 @aj2duncan
Thanks for making all those changes @nelsonroque. I agree entirely with @acolum and I can now too recommend this for publication.
