### 2019-05-28 #1279 @xuanxu
@nelsonroque Some needed minor changes an typos I found in the compiled pdf:

* In the **Statement of need** section: I think this sentence need rewording to make it clearer:
` In raw, form the 2.5 quintillion bytes of raw data generated daily are difficult to interpret -- noisy time-series`
* The Zenodo archive will be linked on the first page of the paper, so we ask that you don't add an additional link in the text or references. Can you edit the *(Roque, 2019)* citation out?
* In the *Functionality* section: There is an orphan closing parentheses in `...separately for each burst)`
* Figures 1 and 2 are misplaced among the text. Please write each of them in its own paragraph.

You can also regenerate the pdf anytime with `@whedon generate pdf` if you want to give the paper further proof-reads.
