### 2019-05-29 39a10b1
updated based on reviews
https://github.com/openjournals/joss-reviews/issues/1279
