+++ paper/paper.md
@@ -29,7 +29,7 @@ affiliations:
 
 # Statement of Need
 
-In today's digital world, data collection and storage costs are quite low. Humans are collectively outputting 2.5 quintillion bytes of data every day; by 2020, each person will generate ~ 1.7 MB every second [@ibmstats].  At this scale, intensive longitudinal data about humans' behavior facilitates new discovery about the patterning of thought and action and potentially better prediction and optimization of health and well-being.  In raw, form the 2.5 quintillion bytes of raw data generated daily are difficult to interpret -- noisy time-series. Extraction of features from the time-series, however, allows:
+In today's digital world, data collection and storage costs are quite low. Humans are collectively outputting 2.5 quintillion bytes of data every day; by 2020, each person will generate ~ 1.7 MB every second [@ibmstats].  At this scale, intensive longitudinal data about humans' behavior facilitates new discovery about the patterning of thought and action and potentially better prediction and optimization of health and well-being.  In raw form, the 2.5 quintillion bytes of data generated daily are noisy time-series and difficult to interpret. Extraction of features from the time-series, however, allows:
 
 1. Researchers to reduce the dimensionality of their time-series data (e.g., reducing millions of time-stamped observations to, for example, summary feature vector of length 100); 
 
@@ -42,7 +42,7 @@ Short data streams are easily summarized using basic features (e.g., mean, stand
 # Summary
 
 ## Functionality
-`tsfeaturex` [@roque_ram_2019] is an R package for automating time series feature extraction, inspired and modeled after the Python package *tsfresh* [@tsfresh; @tsfreshgit].  The R language [@r_cite] allows for an easy to use interface, with the underlying processing speed advantage of C languages (and flexibility to run on the web, with the help of the `shiny` package in R; [@r_shiny]).  The API for `tsfeaturex` is a wrapper for the highly-trafficked `dplyr`[@r_dplyr], mainly to lend on the flexibility of the grammar of data manipulation and shortcuts for non-standard evaluation. The API for `tsfeaturex` was designed to facilitate the extraction of features for any dataset in long format, including grouping of summaries by other factor. For example, if every person in your dataset has 1 observation each day for 8 days, and they do this in two bursts, once every 6 months, you can calculate features of the overall series, 16 observations from both bursts, or separately for each burst). Some features are integrated from other packages, such as: [`e1071`, @r_e1071; `Hmisc`, @r_hmisc; `forecast`, @r_forecast; `zoo`, @r_zoo; `viridis`, @r_viridis; `psych`, @r_psych; `entropy`, @r_entropy; `Langevin`, @Friedrich2011; @Rinn2016].
[…]
