+++ paper/paper.bib
@@ -63,17 +63,6 @@
   year = {2008}
 }
 
-@misc{roque_ram_2019,
-  author    = {Roque, Nelson},
-  title     = {tsfeaturex},
-  version   = {0.3.6},
-  publisher = {Zenodo},
-  month     = {february},
-  year      = {2019},
-  doi = {10.5281/zenodo.2574990},
-  url = {http://dx.doi.org/10.5281/zenodo.2574990}
-}
-
 @Manual{r_cite,
   title = {R: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
