### 2019-05-29 #1279 @nelsonroque
@xuanxu thank you for your feedback. I've addressed points 1-3 and will be testing formatting of Figures 1 / 2  in generated pdf.
