### 2019-05-29 3c692ba
updated -- seeing if shifting heading fixes formatting
https://github.com/openjournals/joss-reviews/issues/1279
