### 2019-05-29 #1279 @xuanxu
> the figures are shifting to new locations

I don't know why that's happening.
Maybe the image is relocated if there is not enough space left in that page? :thinking:
