### 2019-05-29 #1279 @nelsonroque
> > the figures are shifting to new locations
>
> I don't know why that's happening.
> Maybe the image is relocated if there is not enough space left in that page? 🤔

Thank you! I think it is fixed now -- the latest proof places figures centered on a single page and doesn't break up other sections.
