### 2019-05-30 #1279 @xuanxu
@nelsonroque Great! Next steps:
- After the 0.3.6 version there are almost 60 commits with changes made during the review process, can you release a new version of tsfeaturex?
- Then please make a Zenodo archive and report the DOI here
