### 2019-05-30 #1279 @nelsonroque
@xuanxu thank you! I've released version 0.3.10 on Github and created a Zenodo archive:
DOI: 10.5281/zenodo.3235466

@aj2duncan @arfon @labarba thank you all for your feedback along this process!
