### 2019-05-30 #1279 @nelsonroque
@xuanxu I added a .zenodo.json to the root of the repo on Github with updated author / title -- but not seeing this information in the release on Zenodo (0.3.10). I will create a new release now that should have the updated info.
