### 2019-05-30 #1279 @nelsonroque
@xuanxu here is what the .zenodo.json looks like -- not sure where I went wrong:
https://github.com/nelsonroque/tsfeaturex/compare/v.0.3.10...master
