### 2019-05-30 #1279 @xuanxu
@nelsonroque You can edit the metadata (change title and authors) directly in the zenodo website
