### 2019-05-30 #1279 @xuanxu
This is ready!
Pinging EIC for publication: @openjournals/joss-eics
