+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @techreport{ibmstats,
-  author = {IBM Marketing Cloud},
+  author = {{IBM Marketing Cloud}},
   institution = {IBM},
   year={2017},
   title = {{10 Key Marketing Trends for 2017 and Ideas for Exceeding Customer Expectations}},
@@ -7,17 +7,16 @@
 }
 
 @article{tsfresh,
-  title = "Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests (tsfresh: A Python package)",
-  journal = "Neurocomputing",
-  volume = "307",
-  pages = "72 - 77",
-  year = "2018",
-  issn = "0925-2312",
-  doi = "https://doi.org/10.1016/j.neucom.2018.03.067",
-  url = "http://www.sciencedirect.com/science/article/pii/S0925231218304843",
-  author = "Maximilian Christ and Nils Braun and Julius Neuffer and Andreas W. Kempa-Liehr",
-  keywords = "Feature engineering, Time series, Feature extraction, Feature selection, Machine learning",
-  abstract = "Time series feature engineering is a time-consuming process because scientists and engineers have to consider the multifarious algorithms of signal processing and time series analysis for identifying and extracting meaningful features from time series. The Python package tsfresh (Time Series FeatuRe Extraction on basis of Scalable Hypothesis tests) accelerates this process by combining 63 time series characterization methods, which by default compute a total of 794 time series features, with feature selection on basis automatically configured hypothesis tests. By identifying statistically significant time series characteristics in an early stage of the data science process, tsfresh closes feedback loops with domain experts and fosters the development of domain specific features early on. The package implements standard APIs of time series and machine learning libraries (e.g. pandas and scikit-learn) and is designed for both exploratory analyses as well as straightforward integration into operational data science applications."
+  doi = {10.1016/j.neucom.2018.03.067},
+  url = {https://doi.org/10.1016/j.neucom.2018.03.067},
+  year = {2018},
+  month = sep,
+  publisher = {Elsevier {BV}},
+  volume = {307},
+  pages = {72--77},
+  author = {Maximilian Christ and Nils Braun and Julius Neuffer and Andreas W. Kempa-Liehr},
+  title = {Time Series {FeatuRe} Extraction on basis of Scalable Hypothesis tests (tsfresh {\textendash} A Python package)},
+  journal = {Neurocomputing}
 }
 
 @misc{tsfreshgit,
@@ -31,7 +30,7 @@
 
 @misc{tsfreshgitissue,
   author = {blue-yonder},
-  title = {Allow NaN or None values to be passed in, and silently ignored},
+  title = {Allow {NaN} or {None} values to be passed in, and silently ignored},
   year = {2016},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -40,7 +39,7 @@
 
 @misc{findpeaks,
   author = {stas-g},
-  title = {findPeaks},
+  title = {{findPeaks}},
   year = {2017},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -64,7 +63,7 @@
 }
 
 @Manual{r_cite,
-  title = {R: A Language and Environment for Statistical Computing},
+  title = {{R}: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
@@ -73,7 +72,7 @@
 }
 
 @Manual{r_shiny,
-    title = {shiny: Web Application Framework for R},
+    title = {shiny: Web Application Framework for {R}},
[…]
