### 2019-05-30 #1279 @kyleniemeyer
Hi @nelsonroque, just a few minor changes needed for the paper:

- Can you add full affiliation details (e.g., location) as you'd see in other journal articles?
- I submitted a small PR to fix some typographic issues in the paper, if you can merge that: https://github.com/nelsonroque/tsfeaturex/pull/2
- I also submitted a PR to the bib file to fix some things in there: https://github.com/nelsonroque/tsfeaturex/pull/3
