### 2019-05-30 #1279 @nelsonroque
Hi @kyleniemeyer absolutely, all changes have been made!
