+++ README.md
@@ -1,6 +1,7 @@
 # R package: tsfeaturex
 
 [![DOI](https://zenodo.org/badge/170967121.svg)](https://zenodo.org/badge/latestdoi/170967121)
+[![status](http://joss.theoj.org/papers/aa9198d80b72aecc2418ad94f4e7ab1a/status.svg)](http://joss.theoj.org/papers/aa9198d80b72aecc2418ad94f4e7ab1a)
 
 ## Description
 Calculate many features (over 50) of a time series. [Click here to view the full feature list](docs/feature_list.md)
