### 2019-05-30 #1279 @kyleniemeyer
@nelsonroque it look like the affiliation info is still not quite complete—really I'm just looking for city, state, country (like what'd you'd typically see following department/center and university)
