### 2019-05-31 #1279 @nelsonroque
@kyleniemeyer sorry I was adding affiliation on Zenodo side. I have added it to the `paper.md`
