### 2019-05-31 #1279 @kyleniemeyer
Congrats @nelsonroque on your submission's publication in JOSS! Thanks to @acolum and @aj2duncan for reviewing and @xuanxu for editing!
